+++ src/db-candle.interface.ts
@@ -1,0 +1,7 @@
+export interface DbCandle {
+  open_timestamp: Date;
+  open: number;
+  close: number;
+  high: number;
+  low: number;
+}
