+++ src/product-symbols.enum.ts
@@ -1,0 +1,4 @@
+export enum ProductSymbols {
+  ETH = 'ETH',
+  USD = 'USD',
+}
